Move Layout class into main.js
